[fix] Apparently, `ClusterDeployment` must have same name as namespace
This one time, I'm not squashing the fix into the feature commit, because this is a costly finding.

As an educated guess (documented in the comment), there is one or more places in the OpenShift / RHACM code base that assume that the `ClusterDeployment` objects must have the same name as the namespace they live in.

As a matter of fact, this one change causes the “import cluster” clicky feature to start working.