[scaffold] `inventory_manually_created_cluster_name`
Setting this variable in the inventory for a given cluster (on a temporary basis during testing), signals that this cluster is to be created by clicking on the Create button in https://console-openshift-console.apps.mgmt.ocp-test.epfl.ch/multicloud/infrastructure/clusters/managed , rather than through the infrastructure-as-code.

When `inventory_manually_created_cluster_name` is set:
- short-circuit `roles/openshift4-spoke/tasks/spoke-install.yml`
- override `spoke_namespace_in_hub` with the value of `inventory_manually_created_cluster_name`, to work around the fact that the clicky configurator conflates the name of the cluster with the lowest-level component of its domain name (its `cluster_name`, in the parlance of `vars/openshift-vars.yml`).

The rest of the infrastructure-as-code is *not* supposed to know or care about this variable, meaning that one can and should apply “hour-2” configuration, through `kubernetes.core.k8s` or otherwise, regardless of whether the cluster was created manually or automatically.